[test-schools] Refactor TestInstance and instances into utility
